# Cache the DetectableDB scan results