--- a/src/process/index.js
+++ b/src/process/index.js
@@ -12,7 +12,7 @@ import * as Natives from './native/index.js';
 const Native = Natives[process.platform];
 
 
-const timestamps = {}, names = {}, pids = {};
+const timestamps = {}, names = {}, pids = {}, cache = {};
 export default class ProcessServer {
   constructor(handlers) {
     if (!Native) return; // log('unsupported platform:', process.platform);
@@ -36,52 +36,68 @@ export default class ProcessServer {
 
     for (const [ pid, _path, args ] of processes) {
       const path = _path.toLowerCase().replaceAll('\\', '/');
-      const toCompare = [];
-      const splitPath = path.split('/');
-      for (let i = 1; i < splitPath.length; i++) {
-        toCompare.push(splitPath.slice(-i).join('/'));
+      
+      // the cache key includes every dependency of the DetectableDB scan
+      const cacheKey = `${path}\0${args}`;
+      
+      var detected = [];
+      if (cache[cacheKey] !== undefined) {
+        detected = cache[cacheKey].detected;
       }
-
-      for (const p of toCompare.slice()) { // add more possible tweaked paths for less false negatives
-        toCompare.push(p.replace('64', '')); // remove 64bit identifiers-ish
-        toCompare.push(p.replace('.x64', ''));
-        toCompare.push(p.replace('x64', ''));
-        toCompare.push(p.replace('_64', ''));
-      }
-
-      for (const { executables, id, name } of DetectableDB) {
-        if (executables?.some(x => {
-          if (x.is_launcher) return false;
-          if (x.name[0] === '>' ? x.name.substring(1) !== toCompare[0] : !toCompare.some(y => x.name === y)) return false;
-          if (args && x.arguments) return args.join(" ").indexOf(x.arguments) > -1;
-          return true;
-        })) {
-          names[id] = name;
-          pids[id] = pid;
-
-          ids.push(id);
-          if (!timestamps[id]) {
-            log('detected game!', name);
-            timestamps[id] = Date.now();
-          }
-
-          // Resending this on evry scan is intentional, so that in the case that arRPC scans processes before Discord, existing activities will be sent
-          this.handlers.message({
-            socketId: id
-          }, {
-            cmd: 'SET_ACTIVITY',
-            args: {
-              activity: {
-                application_id: id,
-                name,
-                timestamps: {
-                  start: timestamps[id]
-                }
-              },
-              pid
-            }
-          });
+      else {
+        const toCompare = [];
+        const splitPath = path.split('/');
+        for (let i = 1; i < splitPath.length; i++) {
+          toCompare.push(splitPath.slice(-i).join('/'));
         }
+
+        for (const p of toCompare.slice()) { // add more possible tweaked paths for less false negatives
+          toCompare.push(p.replace('64', '')); // remove 64bit identifiers-ish
+          toCompare.push(p.replace('.x64', ''));
+          toCompare.push(p.replace('x64', ''));
+          toCompare.push(p.replace('_64', ''));
+        }
+        
+        for (const { executables, id, name } of DetectableDB) {
+          if (executables?.some(x => {
+            if (x.is_launcher) return false;
+            if (x.name[0] === '>' ? x.name.substring(1) !== toCompare[0] : !toCompare.some(y => x.name === y)) return false;
+            if (args && x.arguments) return args.join(" ").indexOf(x.arguments) > -1;
+            return true;
+          })) {
+            detected.push({ executables, id, name });
+          }
+        }
+        
+        cache[cacheKey] = {"detected": detected};
+      }
+      
+      for (const { executables, id, name } of detected) {
+        names[id] = name;
+        pids[id] = pid;
+
+        ids.push(id);
+        if (!timestamps[id]) {
+          log('detected game!', name);
+          timestamps[id] = Date.now();
+        }
+
+        // Resending this on evry scan is intentional, so that in the case that arRPC scans processes before Discord, existing activities will be sent
+        this.handlers.message({
+          socketId: id
+        }, {
+          cmd: 'SET_ACTIVITY',
+          args: {
+            activity: {
+              application_id: id,
+              name,
+              timestamps: {
+                start: timestamps[id]
+              }
+            },
+            pid
+          }
+        });
       }
     }