Inspired by a performance problem rightfully pointed out in PR #92, I set out to greatly optimize the process scan:

| Scenario              | Time  | Notes                                                                                         |
|-|-|-|
| **`main`** |         |                                                                                               |
| Before                | 🟨 133.7ms |                                                                                |
| After                 | 🟩 2.3ms   | Highly skewed due to what I think are GC spikes, most scans run way below 1ms!|
| **With #92** |         |                                                                                               |
| Before                | 🟥 2699ms  |                                                                                 |
| After                 | 🟩 0.3ms   | NOT faster than `main`, simply no GC spikes happened here                                         |

The change works on my end, but is also provably correct __(assuming `DetectableDB` does not change)__ because:
- the scan results depend on `DetectableDB`'s contents, `args`, and `toCompare`
- `toCompare` is derived from `splitPath`
- `splitPath` is derived from `path`

Notes:
- for the times, I subtracted the time it takes to get processes using `await Native.getProcesses();`. #92 does add functionality to this call but it's negligible (16.3ms -> 18.3ms)
- my system has a Ryzen 9 7940HS, the difference is much more significant on lower end setups where the process scan could have taken more than the 5 second budget and got into a loop of continuously consuming 1 CPU core!
- the results were recorded under no system activity, new processes coming to be will cause spikes but hopefully never worse than redundantly running every single process by `DetectableDB`.

Open questions:
- I'm not sure that the theoretically possible case of multiple `DetectableDB` entries being detected with one process is intentional, but I cache an array of all detected entries to keep this behavior