@Dekomoro 2024-11-18
It appears in combo with 92 the cpu and memory usage goes up but not by much. I do see the memory usage going up at a very slow rate consistently in combination with 92 while idling and faster while multiple different processes are running.
I do think purging unseen processes from the cache may solve this, but I lack the expertise to know for sure.