@hazzuk 2025-04-09
When evaluating this change, I was really blown away by the performance improvements.

Using the logging in `src/process/index.js`, on average I saw a before scan time of `95ms`, and an after of `<0.5ms`. Inline with the original performance testing documented above.

Additionally, running clinic.js `clinic doctor -- node src` gives another view of the improvements to CPU usage, memory stability, and latency.

#### Before
![std-Screenshot 2025-04-09 142712](https://github.com/user-attachments/assets/e20fb7de-8c95-4ed2-a1b3-cd7f85c933c1)

#### After
![cache+std-Screenshot 2025-04-09 142550](https://github.com/user-attachments/assets/ddf30d05-b789-49da-86a1-2cd31747f30e)

Looking forward to seeing this merged, amazing work.